+++ waspc/cli/src/Wasp/Cli/Command/Info.hs
@@ -3,22 +3,16 @@ module Wasp.Cli.Command.Info
   )
 where
 
-import Control.Arrow ()
 import Control.Monad.IO.Class (MonadIO (liftIO))
-import StrongPath (Abs, Dir, Path', fromRelFile)
-import StrongPath.Operations ()
-import System.Directory (getFileSize)
 import qualified Wasp.AppSpec.Valid as ASV
 import Wasp.Cli.Command (Command, require)
-import Wasp.Cli.Command.Common (readWaspCompileInfo)
+import Wasp.Cli.Command.Common (readDirectorySizeMB, readWaspCompileInfo)
 import Wasp.Cli.Command.Compile (analyze)
 import Wasp.Cli.Command.Message (cliSendMessageC)
 import Wasp.Cli.Command.Require.InWaspProject (InWaspProject (InWaspProject))
 import Wasp.Cli.Command.Require.WaspSpecAvailable (WaspSpecAvailable (WaspSpecAvailable))
 import Wasp.Cli.Terminal (title)
 import qualified Wasp.Message as Msg
-import Wasp.Project (WaspProjectDir)
-import qualified Wasp.Util.IO as IOUtil
 import qualified Wasp.Util.Terminal as Term
 
 info :: Command ()
@@ -45,8 +39,3 @@ info = do
 
 printInfo :: String -> String -> String
 printInfo key value = Term.applyStyles [Term.Cyan] key ++ ": " <> Term.applyStyles [Term.White] value
-
-readDirectorySizeMB :: Path' Abs (Dir WaspProjectDir) -> IO String
-readDirectorySizeMB path = (++ " MB") . show . (`div` 1000000) . sum <$> allFileSizes
-  where
-    allFileSizes = IOUtil.listDirectoryDeep path >>= mapM (getFileSize . fromRelFile)